requirements.txt
fastapi
mysql-connector-python
uvicorn[standard]
sqlalchemy
passlib[bcrypt]
pyjwt
psycopg2
faker
faker-commerce